Split ApiServices: extract SourceEnricher for PDB/enrichment orchestration
ApiServices (926 lines) was a god service mixing extraction and enrichment.
Split into two focused classes:

- ApiServices (333 lines): API surface extraction — FindType, ExtractFullApi,
  ResolveForwardedTypes, and the async pipeline methods that orchestrate
  package extraction and type lookup.

- SourceEnricher (605 lines): PDB acquisition, source/doc enrichment —
  AcquirePdbAsync, EnrichTypeWithSourceInfoAsync, batched enrichment,
  repository URL extraction, XML doc fallback, and doc comment merging.

Also extracted ResolvePackageInfo helper to deduplicate package name/version
resolution that was repeated in three enrichment methods.
